dev/core#905 Add contribution recur statuses 'Processing' and 'Failing'
Per dev/core#905 there are some statuses that are useful to record for recurrings that are in addition to those currently in core.
These are

- Processing - when a payment against a contribution recur is actively being processed. This would be used when processing
a referring contribution to alter the status prior to the attempt. Then when the attempt has completed
it would go back to 'In Progress' or whatever  but if it got stuck then it would say 'Processing' - indicating
not to 'just try again'

- Failing - when one or more payemnts against a recurring contribution has failed but the processor has not yet given up.

It would be up to processors to implement these - we are standardising the concepts at this stage - although
we might provide a generic cron that processors can 'opt into
 later