Assign highlighted fields as json
The import map field screen doesn't like the use of implode here
because it has settled on 'join' as the preferred
alias https://www.php.net/manual/en/function.join.php
(probably seen as more user friendly to template developers).

I also don't like the use of join here - because it's really hard to read the line
of code it is used in. Hence I switched to assigning json at the
php layer instead.

Note that the attempt to add colour seems broken & when I checked on 5.70 it seems
to be not working there either - this makes that no better of worse